Move Pkg into project_config module
