+++ src/pkg_version.rs
@@ -2,12 +2,11 @@ use pubgrub::version::{SemanticVersion as SemVer, VersionParseError};
 use serde::{Deserialize, Serialize};
 use serde_json;
 use std::collections::{BTreeMap, BTreeSet};
-use std::fmt;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
 use thiserror::Error;
 
-use crate::project_config::PackageConfig;
+use crate::project_config::{PackageConfig, Pkg, PkgParseError};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(transparent)]
@@ -19,12 +18,6 @@ pub struct Cache {
 pub struct PkgVersion {
     pub author_pkg: Pkg,
     pub version: SemVer,
-}
-
-#[derive(Debug, Clone, Ord, PartialOrd, Eq, PartialEq, Serialize, Deserialize)]
-pub struct Pkg {
-    pub author: String,
-    pub pkg: String,
 }
 
 #[derive(Error, Debug)]
@@ -67,12 +60,6 @@ pub enum PkgVersionParseError {
     VersionParseError(#[from] VersionParseError),
     #[error("failed to parse the package")]
     PkgParseError(#[from] PkgParseError),
-}
-
-#[derive(Error, Debug)]
-pub enum PkgParseError {
-    #[error("no author/package separation found in `{0}`")]
-    NoAuthorSeparator(String),
 }
 
 impl Cache {
@@ -276,37 +263,6 @@ impl PkgVersion {
     }
 }
 
-// Public Pkg methods.
-impl Pkg {
-    pub fn config_path<P: AsRef<Path>>(&self, elm_home: P, elm_version: &str) -> PathBuf {
-        Self::packages_dir(elm_home, elm_version)
-            .join(&self.author)
-            .join(&self.pkg)
-    }
-}
-
-// Private Pkg methods.
-impl Pkg {
-    fn to_url(&self, remote_base_url: &str) -> String {
-        format!("{}/packages/{}/{}", remote_base_url, self.author, self.pkg)
-    }
-
-    fn pubgrub_cache_dir_json<P: AsRef<Path>>(&self, elm_home: P) -> PathBuf {
-        Self::pubgrub_cache_dir(elm_home)
-            .join("elm_json_cache")
-            .join(&self.author)
-            .join(&self.pkg)
-    }
-
-    fn pubgrub_cache_dir<P: AsRef<Path>>(elm_home: P) -> PathBuf {
-        elm_home.as_ref().join("pubgrub")
-    }
-
-    fn packages_dir<P: AsRef<Path>>(elm_home: P, elm_version: &str) -> PathBuf {
-        elm_home.as_ref().join(elm_version).join("packages")
-    }
-}
-
 impl FromStr for PkgVersion {
     type Err = PkgVersionParseError;
     fn from_str(s: &str) -> Result<Self, Self::Err> {
@@ -319,23 +275,5 @@ impl FromStr for PkgVersion {
             author_pkg,
             version,
         })
-    }
-}
-
-impl FromStr for Pkg {
-    type Err = PkgParseError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let author_sep = s
-            .find('/')
-            .ok_or_else(|| PkgParseError::NoAuthorSeparator(s.to_string()))?;
-        let author = s[0..author_sep].to_string();
-        let pkg = s[(author_sep + 1)..].to_string();
-        Ok(Pkg { author, pkg })
-    }
-}
-
-impl fmt::Display for Pkg {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}/{}", &self.author, &self.pkg)
     }
 }
